+++ src/main/java/com/inz/inz/resoruce/reportResource/ReportResource.java
@@ -1,12 +1,10 @@
 package com.inz.inz.resoruce.reportResource;
 
-import com.inz.inz.entity.enums.ReportType;
 import lombok.Data;
 
 @Data
 public class ReportResource extends ReportLight {
 
-    private ReportType reportType;
 
     private String reportDate;
 
@@ -14,7 +12,8 @@ public class ReportResource extends ReportLight {
 
     private Double mark;
 
-
     private Long userId;
+
+    private String video;
 
 }