+++ src/main/java/com/inz/inz/security/adapter/UserAdapter.java
@@ -1,9 +1,7 @@
 package com.inz.inz.security.adapter;
 
 import com.inz.inz.exceptionhandler.DbException;
-import com.inz.inz.security.Resource.UserAuthResoruce;
-import com.inz.inz.security.Resource.UserRank;
-import com.inz.inz.security.Resource.UserResourcePost;
+import com.inz.inz.resoruce.userResource.*;
 import com.inz.inz.security.model.User;
 
 import java.util.List;
@@ -16,4 +14,9 @@ public interface UserAdapter {
     List<UserRank> getRank();
 
     void sendNewPassword(String email) throws DbException;
+
+    void  editUser(EditUserResource user) throws DbException;
+
+    void changePassword(PasswordResourcePost password) throws DbException;
+
 }